+++ src/api/tokens/types.d.ts
@@ -1,0 +1,8 @@
+export interface AccessToken {
+  label: string
+  created: number
+}
+
+export interface AccessTokenSecret {
+  secret: string
+}